+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Changelog
 
+## 1.4.20
+* Bugfix: Extract method did not include parameters from parent method
+* Bugfix: Extract method parameter list was not formatted correctly
+* Bugfix: Extract method was not correctly handling trailing semicolon
+* Bugfix: Name suggestion for Extract Method was suggestion names from outside selection
+
 ## 1.4.19
 * Quickfixes for adding/removing type tags (return types & variable types)
 * Very basic Extract Method support
-* Minor tweaks forIntellij 2023.3 Support 
+* Minor tweaks for Intellij 2023.3 Support 
 
 ## 1.4.18
 * Support for Intellij 2023.3